Preparation of version 1.0 release

+++ dwarfexplorer/QHexView/document/buffer/qhexbuffer.h
@@ -13,18 +13,18 @@ class QHexBuffer : public QObject
         bool isEmpty() const;
 
     public:
-        virtual uchar at(int idx);
-        virtual void replace(int offset, const QByteArray& data);
-        virtual void read(char* data, int size);
-        virtual void read(const QByteArray& ba);
+        virtual uchar at(uint64_t idx);
+        virtual void  replace(uint64_t offset, const QByteArray& data);
+        virtual void  read(char* data, uint64_t size);
+        virtual void  read(const QByteArray& ba);
 
     public:
-        virtual int length() const = 0;
-        virtual void insert(int offset, const QByteArray& data) = 0;
-        virtual void remove(int offset, int length) = 0;
-        virtual QByteArray read(int offset, int length) = 0;
-        virtual void read(QIODevice* iodevice) = 0;
-        virtual void write(QIODevice* iodevice) = 0;
+        virtual uint64_t   length() const = 0;
+        virtual void       insert(uint64_t offset, const QByteArray& data) = 0;
+        virtual void       remove(uint64_t offset, uint64_t length) = 0;
+        virtual QByteArray read(uint64_t offset, uint64_t length) = 0;
+        virtual void       read(QIODevice* iodevice) = 0;
+        virtual void       write(QIODevice* iodevice) = 0;
 };
 
 #endif // QHEXBUFFER_H
