+++ dwarfexplorer/QHexView/document/buffer/qmemoryrefbuffer.h
@@ -9,13 +9,13 @@ class QMemoryRefBuffer : public QHexBuffer
     Q_OBJECT
 
     public:
-        explicit QMemoryRefBuffer(QObject *parent = nullptr);
-        int length() const override;
-        void insert(int, const QByteArray&) override;
-        void remove(int offset, int length) override;
-        QByteArray read(int offset, int length) override;
-        void read(QIODevice* device) override;
-        void write(QIODevice* device) override;
+        explicit   QMemoryRefBuffer(QObject *parent = nullptr);
+        uint64_t   length() const override;
+        void       insert(uint64_t, const QByteArray&) override;
+        void       remove(uint64_t offset, uint64_t length) override;
+        QByteArray read(uint64_t offset, uint64_t length) override;
+        void       read(QIODevice* device) override;
+        void       write(QIODevice* device) override;
 
     private:
         QBuffer* m_buffer;
